refactor(global):
- Rename directories
- Rewrite README.md
- Global-ify starter files
