+++ AloneChat/core/__init__.py
@@ -1,0 +1,4 @@
+from .network.protocol import Message, MessageType
+from .plugin import Plugin, PluginManager
+
+__all__ = ['Message', 'MessageType', 'Plugin', 'PluginManager']
