+++ AloneChat/core/client/__init__.py
@@ -1,0 +1,3 @@
+from .command import CommandSystem
+
+__all__ = ['CommandSystem']
