+++ AloneChat/core/network/__init__.py
@@ -1,0 +1,4 @@
+from .websocket import WebSocketManager
+from .protocol import Message, MessageType
+
+__all__ = ['WebSocketManager', 'Message', 'MessageType']
