+++ AloneChat/core/server/__init__.py
@@ -1,0 +1,2 @@
+# 预留服务端扩展接口
+pass
