+++ AloneChat/plugins/admin/__init__.py
@@ -1,0 +1,3 @@
+from .kick import PluginImpl as KickPlugin
+
+__all__ = ['KickPlugin']
