+++ AloneChat/plugins/encryption/__init__.py
@@ -1,0 +1,3 @@
+from .aes import PluginImpl as AESEncryptor
+
+__all__ = ['AESEncryptor']
