--- Chat_System/core/__init__.py
@@ -1,4 +1,0 @@
-from .network.protocol import Message, MessageType
-from .plugin import Plugin, PluginManager
-
-__all__ = ['Message', 'MessageType', 'Plugin', 'PluginManager']
