--- Chat_System/core/client/__init__.py
@@ -1,3 +1,0 @@
-from .command import CommandSystem
-
-__all__ = ['CommandSystem']
