--- Chat_System/core/network/__init__.py
@@ -1,4 +1,0 @@
-from .websocket import WebSocketManager
-from .protocol import Message, MessageType
-
-__all__ = ['WebSocketManager', 'Message', 'MessageType']
