--- Chat_System/core/server/__init__.py
@@ -1,2 +1,0 @@
-# 预留服务端扩展接口
-pass
