--- Chat_System/plugins/admin/__init__.py
@@ -1,3 +1,0 @@
-from .kick import PluginImpl as KickPlugin
-
-__all__ = ['KickPlugin']
