--- Chat_System/plugins/encryption/__init__.py
@@ -1,3 +1,0 @@
-from .aes import PluginImpl as AESEncryptor
-
-__all__ = ['AESEncryptor']
