+++ server.py
@@ -1,0 +1,5 @@
+from AloneChat.core.network.websocket import WebSocketManager
+
+if __name__ == "__main__":
+    server = WebSocketManager(port=8765)
+    server.start()
